requirements.txt
streamlit
onnxruntime
numpy
opencv-python
Pillow
ultralytics
